Module Lambda_streams_lwt.Async

val first_to_lwt : 'a Lambda_streams.Finite.Async.t -> 'a Lwt.t

Gets the first value from a lambda stream and turns it into a lwt promise

val last_to_lwt : 'a Lambda_streams.Finite.Async.t -> 'a Lwt.t

Gets the last value from a lambda stream and turns it into a lwt promise

val to_lwt_list : 'a Lambda_streams.Finite.Async.t -> 'a list Lwt.t

val from_lwt : 'a Lwt.t -> 'a Lambda_streams.Finite.Async.t

val to_lwt_stream : 'a Lambda_streams.Finite.Async.t -> 'a Lwt_stream.t

val from_lwt_stream : 'a Lwt_stream.t -> 'a Lambda_streams.Finite.Async.t